refactor package location to experimental/kubernetes
