+++ MANIFEST.in
@@ -2,4 +2,4 @@ recursive-include tfx/examples/chicago_taxi_pipeline/data *
 include tfx/orchestration/experimental/interactive/export_templates/*.tmpl
 include tfx/experimental/templates/taxi/.gitignore
 include tfx/experimental/templates/taxi/*.ipynb
-include tfx/extensions/google_cloud_kubernetes/pusher/yaml/*.yaml
+include tfx/extensions/experimental.kubernetes/pusher/yaml/*.yaml
